Move StrahlkorperGr to pointwise functions
