+++ src/Executables/FindHorizons/CMakeLists.txt
@@ -14,6 +14,7 @@ set(LIBS_TO_LINK
   Options
   Parallel
   ParallelInterpolation
+  StrahlkorperGr
   Utilities
   )
