+++ src/PointwiseFunctions/GeneralRelativity/Surfaces/UnitNormalOneForm.hpp
@@ -1,0 +1,36 @@
+// Distributed under the MIT License.
+// See LICENSE.txt for details.
+
+#pragma once
+
+#include "DataStructures/Tensor/TypeAliases.hpp"
+
+/// \cond
+class DataVector;
+namespace gsl {
+template <typename>
+struct not_null;
+}  // namespace gsl
+/// \endcond
+
+namespace StrahlkorperGr {
+/// @{
+/// \ingroup SurfacesGroup
+/// \brief Computes normalized unit normal one-form to a Strahlkorper.
+///
+/// \details The input argument `normal_one_form` \f$n_i\f$ is the
+/// unnormalized surface one-form; it depends on a Strahlkorper but
+/// not on a metric.  The input argument `one_over_one_form_magnitude`
+/// is \f$1/\sqrt{g^{ij}n_i n_j}\f$, which can be computed using (one
+/// over) the `magnitude` function.
+template <typename Frame>
+void unit_normal_one_form(gsl::not_null<tnsr::i<DataVector, 3, Frame>*> result,
+                          const tnsr::i<DataVector, 3, Frame>& normal_one_form,
+                          const DataVector& one_over_one_form_magnitude);
+
+template <typename Frame>
+tnsr::i<DataVector, 3, Frame> unit_normal_one_form(
+    const tnsr::i<DataVector, 3, Frame>& normal_one_form,
+    const DataVector& one_over_one_form_magnitude);
+/// @}
+}  // namespace StrahlkorperGr
